Revert "feat: Добавление Booking и хранение всех данных в бд"
