--- src/main/java/ru/practicum/shareit/ApplicationConfig.java
@@ -1,9 +1,0 @@
-package ru.practicum.shareit;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@Configuration
-@EnableJpaRepositories(basePackages = "ru.practicum.shareit.features")
-public class ApplicationConfig {
-}
