+++ src/main/java/ru/practicum/shareit/booking/Booking.java
@@ -1,0 +1,21 @@
+package ru.practicum.shareit.booking;
+
+import lombok.Data;
+import ru.practicum.shareit.item.model.Item;
+
+import java.util.Date;
+
+/**
+ * TODO Sprint add-bookings.
+ */
+@Data
+public class Booking {
+
+    long bookingId;
+
+    Item item;
+
+    Date toDate;
+
+    Date fromDate;
+}
