+++ src/main/java/ru/practicum/shareit/booking/dto/BookingDto.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.booking.dto;
+
+/**
+ * TODO Sprint add-bookings.
+ */
+public class BookingDto {
+}
