+++ src/main/java/ru/practicum/shareit/exceptions/ErrorHandler.java
@@ -26,9 +26,4 @@ public class ErrorHandler {
         return new ErrorResponse("Сущность не найдена", e.getMessage());
     }
 
-    @ExceptionHandler
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    public ErrorResponse validation(final StatusException e) {
-        return new ErrorResponse("Unknown state: UNSUPPORTED_STATUS", e.getMessage());
-    }
 }
