--- src/main/java/ru/practicum/shareit/exceptions/StatusException.java
@@ -1,8 +1,0 @@
-package ru.practicum.shareit.exceptions;
-
-public class StatusException extends Exception {
-
-    public StatusException(final String message) {
-        super(message);
-    }
-}
