--- src/main/java/ru/practicum/shareit/features/booking/BookingStatus.java
@@ -1,8 +1,0 @@
-package ru.practicum.shareit.features.booking;
-
-public enum BookingStatus {
-    WAITING,
-    APPROVED,
-    REJECTED,
-    CANCELED
-}
