--- src/main/java/ru/practicum/shareit/features/item/ItemRepository.java
@@ -1,8 +1,0 @@
-package ru.practicum.shareit.features.item;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.shareit.features.item.model.Item;
-
-public interface ItemRepository extends JpaRepository<Item, Long> {
-
-}
