--- src/main/java/ru/practicum/shareit/features/item/model/Comment.java
@@ -1,40 +1,0 @@
-package ru.practicum.shareit.features.item.model;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-import ru.practicum.shareit.features.user.model.User;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-@Entity
-@Data
-@Builder
-@Table(name = "comments")
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@AllArgsConstructor
-@NoArgsConstructor
-public class Comment {
-
-    @Id
-    @Column(name = "comment_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
-
-    @Column
-    @NotBlank
-    String text;
-
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "item_id")
-    Item item;
-
-    @NotNull
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    User user;
-
-
-}
