--- src/main/java/ru/practicum/shareit/features/item/model/Item.java
@@ -1,41 +1,0 @@
-package ru.practicum.shareit.features.item.model;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-import ru.practicum.shareit.features.user.model.User;
-
-import javax.persistence.*;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-@Entity
-@Table(name = "items")
-@Builder
-@Data
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@NoArgsConstructor
-@AllArgsConstructor
-public class Item {
-
-    @Id
-    @Column(name = "item_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
-
-    @Column
-    @NotBlank
-    String name;
-
-    @Column
-    @NotBlank
-    String description;
-
-    @Column
-    @NotNull
-    Boolean available;
-
-    @ManyToOne
-    @NotNull
-    User owner;
-
-}
