--- src/main/java/ru/practicum/shareit/features/item/model/ItemShortIdWithName.java
@@ -1,16 +1,0 @@
-package ru.practicum.shareit.features.item.model;
-
-import lombok.*;
-import lombok.experimental.FieldDefaults;
-
-@Data
-@Builder
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@NoArgsConstructor
-@AllArgsConstructor
-public class ItemShortIdWithName {
-
-    Long id;
-
-    String name;
-}
