--- src/main/java/ru/practicum/shareit/features/request/ItemRequest.java
@@ -1,4 +1,0 @@
-package ru.practicum.shareit.features.request;
-
-public class ItemRequest {
-}
