--- src/main/java/ru/practicum/shareit/features/request/ItemRequestController.java
@@ -1,10 +1,0 @@
-package ru.practicum.shareit.features.request;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-
-@RestController
-@RequestMapping(path = "/requests")
-public class ItemRequestController {
-}
