--- src/main/java/ru/practicum/shareit/features/request/dto/ItemRequestDto.java
@@ -1,4 +1,0 @@
-package ru.practicum.shareit.features.request.dto;
-
-public class ItemRequestDto {
-}
