--- src/main/java/ru/practicum/shareit/features/user/model/User.java
@@ -1,31 +1,0 @@
-package ru.practicum.shareit.features.user.model;
-
-import lombok.AccessLevel;
-import lombok.Data;
-import lombok.experimental.FieldDefaults;
-import ru.practicum.shareit.validations.PatchValidationGroup;
-
-import javax.persistence.*;
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotBlank;
-
-@Entity
-@Table(name = "users")
-@Data
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class User {
-
-    @Id
-    @Column(name = "user_id")
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
-
-    @Column(unique = true)
-    @Email(regexp = "^(?=.{1,64}@)[A-Za-z0-9_-]+(\\.[A-Za-z0-9_-]+)*@"
-            + "[^-][A-Za-z0-9-]+(\\.[A-Za-z0-9-]+)*(\\.[A-Za-z]{2,})$")
-    String email;
-
-    @Column
-    @NotBlank(groups = PatchValidationGroup.NameNotBlank.class)
-    String name;
-}
