--- src/main/java/ru/practicum/shareit/features/user/model/UserDto.java
@@ -1,17 +1,0 @@
-package ru.practicum.shareit.features.user.model;
-
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Data;
-import lombok.experimental.FieldDefaults;
-
-import javax.validation.constraints.NotNull;
-
-@Data
-@Builder
-@FieldDefaults(level = AccessLevel.PRIVATE)
-public class UserDto {
-
-    @NotNull
-    Long id;
-}
