+++ src/main/java/ru/practicum/shareit/request/dto/ItemRequestDto.java
@@ -1,0 +1,7 @@
+package ru.practicum.shareit.request.dto;
+
+/**
+ * TODO Sprint add-item-requests.
+ */
+public class ItemRequestDto {
+}
