+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -1,0 +1,13 @@
+package ru.practicum.shareit.user.model;
+
+import lombok.Data;
+
+@Data
+public class User {
+
+    public Long id;
+
+    public String email;
+
+    public String name;
+}
