--- src/main/java/ru/practicum/shareit/validations/PatchValidationGroup.java
@@ -1,12 +1,0 @@
-package ru.practicum.shareit.validations;
-
-public interface PatchValidationGroup {
-    interface NameNotBlank {
-    }
-
-    interface DescriptionNotBlank {
-    }
-
-    interface AvailableNotBlank {
-    }
-}
