+++ src/test/java/ru/practicum/shareit/ShareItTests.java
@@ -8,7 +8,6 @@ class ShareItTests {
 
 	@Test
 	void contextLoads() {
-
 	}
 
 }
